Fix compiler warning C4267: conversion from 'size_t' to 'DWORD', possible loss of data Fix compiler warning C4100: unreferenced formal parameter
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@6681 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/CtrlHomePanel.cpp
@@ -756,7 +756,7 @@ void CHomeLibraryBox::Update()
 		}
 	}
 	
-	DWORD nHashing = LibraryBuilder.GetRemaining();
+	size_t nHashing = LibraryBuilder.GetRemaining();
 	
 	if ( nHashing > 0 )
 	{
